modificaciones

+++ src/modules/products/products.model.js
@@ -5,7 +5,7 @@ const Product = sequelize.define('products', {
 	id_product: { type: Sequelize.INTEGER, primaryKey: true, autoIncrement: true },
 	reference: { type: Sequelize.STRING(40), unique: true },
 	name: Sequelize.STRING(50),
-	price: Sequelize.DOUBLE,
+	price: Sequelize.DOUBLE(8,2),
 	id_category: { type: Sequelize.INTEGER, references: { model: 'categories', key: 'id_category' } },
 	id_unit: { type: Sequelize.INTEGER, references: { model: 'units', key: 'id_unit' } },
 	//id_trademark: { type: Sequelize.INTEGER, references: { model: 'trademarks', key: 'id_trademark' } },
